+++ mobile/cmp/form/index.js
@@ -7,8 +7,10 @@
 
 export * from './FormField';
 export * from './Label';
+export * from './inputs/Checkbox';
 export * from './inputs/NumberInput';
 export * from './inputs/SearchInput';
+export * from './inputs/Select';
+export * from './inputs/SwitchInput';
 export * from './inputs/TextInput';
 export * from './inputs/TextArea';
-export * from './inputs/Select';